refactor: 피드백 2차 수정 (getChartDataSet)
